Add DTOs for TaskList and TaskItem (#3)
This PR decouples the database entities from the API layer entities that clients interact with. It accomplishes this by mapping the db entities to DTO records. In this way, clients only interact with our DTO models, not the core (internal) database ones. This makes it less likely for a dev to introduce breaking changes to the API's inputs/outputs. I created a dedicated class to do the mappings so that the mapping logic is centralized, as it would otherwise get duplicated in many places.