Track dates that data is created, updated, and deleted (#5)
This PR adds two new fields to our entities: CreatedAt and UpdatedAt. These go along with the existing DeletedAt field to collectively provide a basic audit history of when data is changed. Given that we want to track this on all entities, I've moved the fields to an abstract parent class that all entities inherit from. This has the very attractive side-effect of allowing us to centralize the management of those fields, thanks to an override of SaveChangesAsync, so clients merely have to add/update/delete data, and the fields are set automatically.  In the future, when we have users, we can additionally track who made the changes.