Add initial React UI with Tailwind CSS setup and basic task list (#7)
Sets up a basic UI for viewing task lists and the tasks within them. Tasks lists can be created, and the tasks within them can be added and their completion can be toggled. We're using React together with Tailwind CSS for styling. Obviously, more features need to be built, but this at least gives us a solid platform to build from. Error and loading states could also be added, but let's leave those for a future PR so as to keep this one to a manageable size.