Add a dark mode toggle (#17)
Introduces a control in the upper right of the page to toggle between dark and light mode. This is accomplished using the built in classes from tailwind, plus a new ThemeToggle component that places/removes the "dark" class on the document element. I decided to just use an icon, without text, so as to not clutter the UI with text on a button that will seldom be used.