Allow users to mark tasks as urgent (#26)
This PR introduces support for marking individual tasks as urgent. A new PriorityLevel enum was added (values are Normal and Fire, existing tasks are of course Normal), along with UI and API updates to toggle it from the task row. Consistent with the new "fire" theme, urgent tasks are shown with a fire icon.

As part of this change, I also did some refactoring of the API: I removed the general-purpose "task update" endpoint in favor of individual patch endpoints. I found that the complexity of the update endpoint was getting out of hand otherwise, as we were trying to support optional arguments in the DTO and only update the supplied fields. As the model grows, that approach isn't maintainable. Meanwhile, the individual patch endpoints are short and simple, easy to test and easy to maintain.