Improve React Query Usage  (#27)
This PR makes better use of React Query v5, which got rid of the old onSuccess and onError callbacks in favor of mutation lifecycle methods like onMutate, onSettled, and the new meta messaging pattern. This gives us more consistency in the code and better toast messages in the UI. As part of these changes, I've also moved the toast rendering to its own component, and I've made a small library helper to manage its state.