Resolve client error generated when Draggable can't find drag handles (#30)
This problem was generating a client side error for each list when the user visits the Burning Tasks page. It was happening because the Draggable component was looking for a drag handle and couldn't find one, since there isn't one in that view. These updates instruct draggable to disable itself when in that view, resolving the errors.